Support Oryx 10GigE camera (#18)
This adds support for an Oryx 10GigE (ORX-10GS-51S5M-C) camera.

The underlying implementation is shared with the previously support
Blackfly camera. The advantage of this approach is that there is less
code to maintain and supporting other Spinnaker based cameras could be
little to no work.

The disadvantage of this approach is that the implementation related to
triggers is complicated. Alternatively, we could consider completely
splitting at least part of the implementation and using different parts
of it explicitly based on the camera model.

This also adds tests for the Oryx camera that mirror those for the
Blackfly. A test for `Camera::get_meta` was also added for both cameras.
I also added a ctest setup fixture to ensure that the packet size is set
to 1000B for the Oryx camera.

Closes #3

---------

Co-authored-by: Alan Liddell <alan.c.liddell@gmail.com>