restore original trigger mode when stopping (#121)
This resolves a problem encountered while trying to use software
triggering. Configuring initially enables the software trigger, setting
`TriggerMode` to on. @nenada encountered this while testing on
acquire-micromanager.

When stopping, `TriggerMode` was set to `off`.

Later when `get()` queries the trigger state, it finds the trigger mode
off and returns a configuration with all triggering disabled.

This PR, resolves the issue by changing the logic in `stop()`. The mode
is toggled off, and then restored to it's original value.